@@ -40,7 +40,7 @@
 #define INCLUDE_OKVIS_CERES_HOMOGENEOUSPOINTPARAMETERBLOCK_HPP_
 
 #include <Eigen/Core>
-#include <okvis/ceres/HomogeneousPointLocalParameterization.hpp>
+#include <okvis/ceres/HomogeneousPointManifold.hpp>
 #include <okvis/ceres/ParameterBlockSized.hpp>
 #include <string>
 
@@ -111,14 +111,14 @@ class HomogeneousPointParameterBlock : public ParameterBlockSized<4, 3, Eigen::V
   /// @param[in] Delta_Chi Perturbation.
   /// @param[out] x0_plus_Delta Perturbed x.
   virtual void plus(const double* x0, const double* Delta_Chi, double* x0_plus_Delta) const {
-    HomogeneousPointLocalParameterization::plus(x0, Delta_Chi, x0_plus_Delta);
+    HomogeneousPointManifold::plus(x0, Delta_Chi, x0_plus_Delta);
   }
 
   /// \brief The jacobian of Plus(x, delta) w.r.t delta at delta = 0.
   /// @param[in] x0 Variable.
   /// @param[out] jacobian The Jacobian.
   virtual void plusJacobian(const double* x0, double* jacobian) const {
-    HomogeneousPointLocalParameterization::plusJacobian(x0, jacobian);
+    HomogeneousPointManifold::plusJacobian(x0, jacobian);
   }
 
   // Delta_Chi=x0_plus_Delta[-]x0
@@ -127,8 +127,8 @@ class HomogeneousPointParameterBlock : public ParameterBlockSized<4, 3, Eigen::V
   /// @param[in] x0_plus_Delta Perturbed variable.
   /// @param[out] Delta_Chi Minimal difference.
   /// \return True on success.
-  virtual void minus(const double* x0, const double* x0_plus_Delta, double* Delta_Chi) const {
-    HomogeneousPointLocalParameterization::minus(x0, x0_plus_Delta, Delta_Chi);
+  virtual void minus(const double* x0_plus_Delta, const double* x0, double* Delta_Chi) const {
+    HomogeneousPointManifold::minus(x0_plus_Delta, x0, Delta_Chi);
   }
 
   /// \brief Computes the Jacobian from minimal space to naively overparameterised space as used by ceres.
@@ -136,7 +136,7 @@ class HomogeneousPointParameterBlock : public ParameterBlockSized<4, 3, Eigen::V
   /// @param[out] jacobian the Jacobian (dimension minDim x dim).
   /// \return True on success.
   virtual void liftJacobian(const double* x0, double* jacobian) const {
-    HomogeneousPointLocalParameterization::liftJacobian(x0, jacobian);
+    HomogeneousPointManifold::liftJacobian(x0, jacobian);
   }
 
   /// @brief Return parameter block type as string
