@@ -163,7 +163,7 @@ bool ReprojectionError<GEOMETRY_T>::EvaluateWithMinimalJacobians(double const* c
 
       // pseudo inverse of the local parametrization Jacobian:
       Eigen::Matrix<double, 6, 7, Eigen::RowMajor> J_lift;
-      PoseLocalParameterization::liftJacobian(parameters[0], J_lift.data());
+      PoseManifold::liftJacobian(parameters[0], J_lift.data());
 
       // hallucinate Jacobian w.r.t. state
       Eigen::Map<Eigen::Matrix<double, 2, 7, Eigen::RowMajor> > J0(jacobians[0]);
@@ -209,7 +209,7 @@ bool ReprojectionError<GEOMETRY_T>::EvaluateWithMinimalJacobians(double const* c
 
       // pseudo inverse of the local parametrization Jacobian:
       Eigen::Matrix<double, 6, 7, Eigen::RowMajor> J_lift;
-      PoseLocalParameterization::liftJacobian(parameters[2], J_lift.data());
+      PoseManifold::liftJacobian(parameters[2], J_lift.data());
 
       // hallucinate Jacobian w.r.t. state
       Eigen::Map<Eigen::Matrix<double, 2, 7, Eigen::RowMajor> > J2(jacobians[2]);
