@@ -31,13 +31,13 @@
  *********************************************************************************/
 
 /**
- * @file PoseLocalParameterization.cpp
- * @brief Source file for the PoseLocalParameterization class.
+ * @file PoseManifold.cpp
+ * @brief Source file for the PoseManifold class.
  * @author Stefan Leutenegger
  */
 
 #include <okvis/assert_macros.hpp>
-#include <okvis/ceres/PoseLocalParameterization.hpp>
+#include <okvis/ceres/PoseManifold.hpp>
 #include <okvis/kinematics/Transformation.hpp>
 #include <okvis/kinematics/operators.hpp>
 
@@ -49,15 +49,15 @@ namespace ceres {
 // Generalization of the addition operation,
 //        x_plus_delta = Plus(x, delta)
 //        with the condition that Plus(x, 0) = x.
-bool PoseLocalParameterization::Plus(const double* x, const double* delta, double* x_plus_delta) const {
+bool PoseManifold::Plus(const double* x, const double* delta, double* x_plus_delta) const {
   return plus(x, delta, x_plus_delta);
 }
 
 // Generalization of the addition operation,
 //        x_plus_delta = Plus(x, delta)
 //        with the condition that Plus(x, 0) = x.
-bool PoseLocalParameterization::plus(const double* x, const double* delta, double* x_plus_delta) {
-  Eigen::Map<const Eigen::Matrix<double, 6, 1> > delta_(delta);
+bool PoseManifold::plus(const double* x, const double* delta, double* x_plus_delta) {
+  Eigen::Map<const Eigen::Matrix<double, 6, 1>> delta_(delta);
 
   // transform to okvis::kinematics framework
   okvis::kinematics::Transformation T(Eigen::Vector3d(x[0], x[1], x[2]), Eigen::Quaterniond(x[6], x[3], x[4], x[5]));
@@ -82,17 +82,15 @@ bool PoseLocalParameterization::plus(const double* x, const double* delta, doubl
 }
 
 // Computes the minimal difference between a variable x and a perturbed variable x_plus_delta.
-bool PoseLocalParameterization::Minus(const double* x, const double* x_plus_delta, double* delta) const {
-  return minus(x, x_plus_delta, delta);
+bool PoseManifold::Minus(const double* x_plus_delta, const double* x, double* delta) const {
+  return minus(x_plus_delta, x, delta);
 }
 
 // Computes the Jacobian from minimal space to naively overparameterised space as used by ceres.
-bool PoseLocalParameterization::ComputeLiftJacobian(const double* x, double* jacobian) const {
-  return liftJacobian(x, jacobian);
-}
+bool PoseManifold::ComputeLiftJacobian(const double* x, double* jacobian) const { return liftJacobian(x, jacobian); }
 
 // Computes the minimal difference between a variable x and a perturbed variable x_plus_delta.
-bool PoseLocalParameterization::minus(const double* x, const double* x_plus_delta, double* delta) {
+bool PoseManifold::minus(const double* x_plus_delta, const double* x, double* delta) {
   delta[0] = x_plus_delta[0] - x[0];
   delta[1] = x_plus_delta[1] - x[1];
   delta[2] = x_plus_delta[2] - x[2];
@@ -104,17 +102,31 @@ bool PoseLocalParameterization::minus(const double* x, const double* x_plus_delt
 }
 
 // The jacobian of Plus(x, delta) w.r.t delta at delta = 0.
-bool PoseLocalParameterization::plusJacobian(const double* x, double* jacobian) {
-  Eigen::Map<Eigen::Matrix<double, 7, 6, Eigen::RowMajor> > Jp(jacobian);
+bool PoseManifold::plusJacobian(const double* x, double* jacobian) {
+  Eigen::Map<Eigen::Matrix<double, 7, 6, Eigen::RowMajor>> Jp(jacobian);
   okvis::kinematics::Transformation T(Eigen::Vector3d(x[0], x[1], x[2]), Eigen::Quaterniond(x[6], x[3], x[4], x[5]));
   T.oplusJacobian(Jp);
 
   return true;
 }
 
+// Compute the derivative of Minus(y, x) w.r.t y at y = x
+bool PoseManifold::minusJacobian(const double* x, double* jacobian) {
+  const Eigen::Quaterniond quat(x[6], x[3], x[4], x[5]);
+
+  Eigen::Map<Eigen::Matrix<double, 6, 7, Eigen::RowMajor>> J(jacobian);
+  J.setZero();
+  J.block<3, 3>(0, 0).setIdentity();
+  J.block<3, 4>(3, 3) = 2.0 * okvis::kinematics::plus(quat).topLeftCorner<3, 4>();
+  // TODO(bjoshi): Not sure why the last column is coming negative
+  J.col(6) = -1.0 * J.col(6);
+
+  return true;
+}
+
 // Computes the Jacobian from minimal space to naively overparameterised space as used by ceres.
-bool PoseLocalParameterization::liftJacobian(const double* x, double* jacobian) {
-  Eigen::Map<Eigen::Matrix<double, 6, 7, Eigen::RowMajor> > J_lift(jacobian);
+bool PoseManifold::liftJacobian(const double* x, double* jacobian) {
+  Eigen::Map<Eigen::Matrix<double, 6, 7, Eigen::RowMajor>> J_lift(jacobian);
   const Eigen::Quaterniond q_inv(x[6], -x[3], -x[4], -x[5]);
   J_lift.setZero();
   J_lift.topLeftCorner<3, 3>().setIdentity();
@@ -128,14 +140,15 @@ bool PoseLocalParameterization::liftJacobian(const double* x, double* jacobian) 
 }
 
 // The jacobian of Plus(x, delta) w.r.t delta at delta = 0.
-bool PoseLocalParameterization::ComputeJacobian(const double* x, double* jacobian) const {
-  return plusJacobian(x, jacobian);
-}
+bool PoseManifold::PlusJacobian(const double* x, double* jacobian) const { return plusJacobian(x, jacobian); }
 
-bool PoseLocalParameterization::VerifyJacobianNumDiff(const double* x, double* jacobian, double* jacobianNumDiff) {
+// The Jacobian of Minus(y, x) w.r.t y at y = x
+bool PoseManifold::MinusJacobian(const double* x, double* jacobian) const { return minusJacobian(x, jacobian); }
+
+bool PoseManifold::VerifyJacobianNumDiff(const double* x, double* jacobian, double* jacobianNumDiff) {
   plusJacobian(x, jacobian);
-  Eigen::Map<Eigen::Matrix<double, 7, 6, Eigen::RowMajor> > Jp(jacobian);
-  Eigen::Map<Eigen::Matrix<double, 7, 6, Eigen::RowMajor> > Jpn(jacobianNumDiff);
+  Eigen::Map<Eigen::Matrix<double, 7, 6, Eigen::RowMajor>> Jp(jacobian);
+  Eigen::Map<Eigen::Matrix<double, 7, 6, Eigen::RowMajor>> Jpn(jacobianNumDiff);
   double dx = 1e-9;
   Eigen::Matrix<double, 7, 1> xp;
   Eigen::Matrix<double, 7, 1> xm;
@@ -157,7 +170,7 @@ bool PoseLocalParameterization::VerifyJacobianNumDiff(const double* x, double* j
 // Generalization of the addition operation,
 //        x_plus_delta = Plus(x, delta)
 //        with the condition that Plus(x, 0) = x.
-bool PoseLocalParameterization3d::Plus(const double* x, const double* delta, double* x_plus_delta) const {
+bool PoseManifold3d::Plus(const double* x, const double* delta, double* x_plus_delta) const {
   Eigen::Matrix<double, 6, 1> delta_;
   delta_.setZero();
   delta_[3] = delta[0];
@@ -185,7 +198,12 @@ bool PoseLocalParameterization3d::Plus(const double* x, const double* delta, dou
 }
 
 // Computes the minimal difference between a variable x and a perturbed variable x_plus_delta.
-bool PoseLocalParameterization3d::Minus(const double* x, const double* x_plus_delta, double* delta) const {
+bool PoseManifold3d::Minus(const double* x_plus_delta, const double* x, double* delta) const {
+  return minus(x_plus_delta, x, delta);
+}
+
+// Computes the minimal difference between a variable x and a perturbed variable x_plus_delta.
+bool PoseManifold3d::minus(const double* x_plus_delta, const double* x, double* delta) {
   const Eigen::Quaterniond q_plus_delta_(x_plus_delta[6], x_plus_delta[3], x_plus_delta[4], x_plus_delta[5]);
   const Eigen::Quaterniond q_(x[6], x[3], x[4], x[5]);
   Eigen::Vector3d delta_q_;
@@ -199,13 +217,11 @@ bool PoseLocalParameterization3d::Minus(const double* x, const double* x_plus_de
 }
 
 // Computes the Jacobian from minimal space to naively overparameterised space as used by ceres.
-bool PoseLocalParameterization3d::ComputeLiftJacobian(const double* x, double* jacobian) const {
-  return liftJacobian(x, jacobian);
-}
+bool PoseManifold3d::ComputeLiftJacobian(const double* x, double* jacobian) const { return liftJacobian(x, jacobian); }
 
 // The jacobian of Plus(x, delta) w.r.t delta at delta = 0.
-bool PoseLocalParameterization3d::plusJacobian(const double* x, double* jacobian) {
-  Eigen::Map<Eigen::Matrix<double, 7, 3, Eigen::RowMajor> > Jp(jacobian);
+bool PoseManifold3d::plusJacobian(const double* x, double* jacobian) {
+  Eigen::Map<Eigen::Matrix<double, 7, 3, Eigen::RowMajor>> Jp(jacobian);
   Jp.setZero();
 
   // rotation: dq*q=qoplus(q)*dq=qoplus(q)*1/2*[I_3;0]*dalpha
@@ -218,9 +234,22 @@ bool PoseLocalParameterization3d::plusJacobian(const double* x, double* jacobian
   return true;
 }
 
+// Compute the derivative of Minus(y, x) w.r.t y at y = x
+bool PoseManifold3d::minusJacobian(const double* x, double* jacobian) {
+  const Eigen::Quaterniond quat(x[6], x[3], x[4], x[5]);
+
+  Eigen::Map<Eigen::Matrix<double, 3, 7, Eigen::RowMajor>> J(jacobian);
+  J.setZero();
+  J.block<3, 4>(0, 3) = 2.0 * okvis::kinematics::plus(quat).topLeftCorner<3, 4>();
+  // TODO(bjoshi): Not sure why the last column is coming negative
+  J.col(6) = -1.0 * J.col(6);
+
+  return true;
+}
+
 // Computes the Jacobian from minimal space to naively overparameterised space as used by ceres.
-bool PoseLocalParameterization3d::liftJacobian(const double* x, double* jacobian) {
-  Eigen::Map<Eigen::Matrix<double, 3, 7, Eigen::RowMajor> > J_lift(jacobian);
+bool PoseManifold3d::liftJacobian(const double* x, double* jacobian) {
+  Eigen::Map<Eigen::Matrix<double, 3, 7, Eigen::RowMajor>> J_lift(jacobian);
   J_lift.setZero();
 
   const Eigen::Quaterniond q_inv(x[6], -x[3], -x[4], -x[5]);
@@ -236,14 +265,15 @@ bool PoseLocalParameterization3d::liftJacobian(const double* x, double* jacobian
 }
 
 // The jacobian of Plus(x, delta) w.r.t delta at delta = 0.
-bool PoseLocalParameterization3d::ComputeJacobian(const double* x, double* jacobian) const {
-  return plusJacobian(x, jacobian);
-}
+bool PoseManifold3d::PlusJacobian(const double* x, double* jacobian) const { return plusJacobian(x, jacobian); }
+
+// Compute the derivative of Minus(y, x) w.r.t y at y = x
+bool PoseManifold3d::MinusJacobian(const double* x, double* jacobian) const { return minusJacobian(x, jacobian); }
 
 // Generalization of the addition operation,
 //        x_plus_delta = Plus(x, delta)
 //        with the condition that Plus(x, 0) = x.
-bool PoseLocalParameterization4d::Plus(const double* x, const double* delta, double* x_plus_delta) const {
+bool PoseManifold4d::Plus(const double* x, const double* delta, double* x_plus_delta) const {
   Eigen::Matrix<double, 6, 1> delta_;
   delta_.setZero();
   delta_[0] = delta[0];
@@ -272,7 +302,12 @@ bool PoseLocalParameterization4d::Plus(const double* x, const double* delta, dou
 }
 
 // Computes the minimal difference between a variable x and a perturbed variable x_plus_delta.
-bool PoseLocalParameterization4d::Minus(const double* x, const double* x_plus_delta, double* delta) const {
+bool PoseManifold4d::Minus(const double* x_plus_delta, const double* x, double* delta) const {
+  return minus(x_plus_delta, x, delta);
+}
+
+// Computes the minimal difference between a variable x and a perturbed variable x_plus_delta.
+bool PoseManifold4d::minus(const double* x_plus_delta, const double* x, double* delta) {
   delta[0] = x_plus_delta[0] - x[0];
   delta[1] = x_plus_delta[1] - x[1];
   delta[2] = x_plus_delta[2] - x[2];
@@ -284,13 +319,11 @@ bool PoseLocalParameterization4d::Minus(const double* x, const double* x_plus_de
 }
 
 // Computes the Jacobian from minimal space to naively overparameterised space as used by ceres.
-bool PoseLocalParameterization4d::ComputeLiftJacobian(const double* x, double* jacobian) const {
-  return liftJacobian(x, jacobian);
-}
+bool PoseManifold4d::ComputeLiftJacobian(const double* x, double* jacobian) const { return liftJacobian(x, jacobian); }
 
 // The jacobian of Plus(x, delta) w.r.t delta at delta = 0.
-bool PoseLocalParameterization4d::plusJacobian(const double* x, double* jacobian) {
-  Eigen::Map<Eigen::Matrix<double, 7, 4, Eigen::RowMajor> > Jp(jacobian);
+bool PoseManifold4d::plusJacobian(const double* x, double* jacobian) {
+  Eigen::Map<Eigen::Matrix<double, 7, 4, Eigen::RowMajor>> Jp(jacobian);
   Eigen::Matrix<double, 7, 6, Eigen::RowMajor> Jp_full;
   okvis::kinematics::Transformation T(Eigen::Vector3d(x[0], x[1], x[2]), Eigen::Quaterniond(x[6], x[3], x[4], x[5]));
   T.oplusJacobian(Jp_full);
@@ -299,9 +332,22 @@ bool PoseLocalParameterization4d::plusJacobian(const double* x, double* jacobian
   return true;
 }
 
+// Compute the derivative of Minus(y, x) w.r.t y at y = x
+bool PoseManifold4d::minusJacobian(const double* x, double* jacobian) {
+  const Eigen::Quaterniond quat(x[6], x[3], x[4], x[5]);
+
+  Eigen::Map<Eigen::Matrix<double, 4, 7, Eigen::RowMajor>> J(jacobian);
+
+  J.setZero();
+  J.block<3, 3>(0, 0).setIdentity();
+  J.block<1, 4>(3, 3) = 2.0 * okvis::kinematics::plus(quat).block<1, 4>(2, 0);
+  J.col(6) = -1.0 * J.col(6);
+  return true;
+}
+
 // Computes the Jacobian from minimal space to naively overparameterised space as used by ceres.
-bool PoseLocalParameterization4d::liftJacobian(const double* x, double* jacobian) {
-  Eigen::Map<Eigen::Matrix<double, 4, 7, Eigen::RowMajor> > J_lift(jacobian);
+bool PoseManifold4d::liftJacobian(const double* x, double* jacobian) {
+  Eigen::Map<Eigen::Matrix<double, 4, 7, Eigen::RowMajor>> J_lift(jacobian);
   const Eigen::Quaterniond q_inv(x[6], -x[3], -x[4], -x[5]);
   J_lift.setZero();
   J_lift.topLeftCorner<3, 3>().setIdentity();
@@ -315,14 +361,15 @@ bool PoseLocalParameterization4d::liftJacobian(const double* x, double* jacobian
 }
 
 // The jacobian of Plus(x, delta) w.r.t delta at delta = 0.
-bool PoseLocalParameterization4d::ComputeJacobian(const double* x, double* jacobian) const {
-  return plusJacobian(x, jacobian);
-}
+bool PoseManifold4d::PlusJacobian(const double* x, double* jacobian) const { return plusJacobian(x, jacobian); }
+
+// Compute the derivative of Minus(y, x) w.r.t y at y = x
+bool PoseManifold4d::MinusJacobian(const double* x, double* jacobian) const { return minusJacobian(x, jacobian); }
 
 // Generalization of the addition operation,
 //        x_plus_delta = Plus(x, delta)
 //        with the condition that Plus(x, 0) = x.
-bool PoseLocalParameterization2d::Plus(const double* x, const double* delta, double* x_plus_delta) const {
+bool PoseManifold2d::Plus(const double* x, const double* delta, double* x_plus_delta) const {
   Eigen::Matrix<double, 6, 1> delta_;
   delta_.setZero();
   delta_[3] = delta[0];
@@ -349,7 +396,12 @@ bool PoseLocalParameterization2d::Plus(const double* x, const double* delta, dou
 }
 
 // Computes the minimal difference between a variable x and a perturbed variable x_plus_delta.
-bool PoseLocalParameterization2d::Minus(const double* x, const double* x_plus_delta, double* delta) const {
+bool PoseManifold2d::Minus(const double* x_plus_delta, const double* x, double* delta) const {
+  return minus(x_plus_delta, x, delta);
+}
+
+// Computes the minimal difference between a variable x and a perturbed variable x_plus_delta.
+bool PoseManifold2d::minus(const double* x_plus_delta, const double* x, double* delta) {
   const Eigen::Quaterniond q_plus_delta_(x_plus_delta[6], x_plus_delta[3], x_plus_delta[4], x_plus_delta[5]);
   const Eigen::Quaterniond q_(x[6], x[3], x[4], x[5]);
   Eigen::Vector3d delta_q_;
@@ -362,13 +414,11 @@ bool PoseLocalParameterization2d::Minus(const double* x, const double* x_plus_de
 }
 
 // Computes the Jacobian from minimal space to naively overparameterised space as used by ceres.
-bool PoseLocalParameterization2d::ComputeLiftJacobian(const double* x, double* jacobian) const {
-  return liftJacobian(x, jacobian);
-}
+bool PoseManifold2d::ComputeLiftJacobian(const double* x, double* jacobian) const { return liftJacobian(x, jacobian); }
 
 // The jacobian of Plus(x, delta) w.r.t delta at delta = 0.
-bool PoseLocalParameterization2d::plusJacobian(const double* x, double* jacobian) {
-  Eigen::Map<Eigen::Matrix<double, 7, 2, Eigen::RowMajor> > Jp(jacobian);
+bool PoseManifold2d::plusJacobian(const double* x, double* jacobian) {
+  Eigen::Map<Eigen::Matrix<double, 7, 2, Eigen::RowMajor>> Jp(jacobian);
   Jp.setZero();
 
   // rotation: dq*q=qoplus(q)*dq=qoplus(q)*1/2*[I_3;0]*dalpha
@@ -381,9 +431,21 @@ bool PoseLocalParameterization2d::plusJacobian(const double* x, double* jacobian
   return true;
 }
 
+// Compute the derivative of Minus(y, x) w.r.t y at y = x
+bool PoseManifold2d::minusJacobian(const double* x, double* jacobian) {
+  const Eigen::Quaterniond quat(x[6], x[3], x[4], x[5]);
+
+  Eigen::Map<Eigen::Matrix<double, 2, 7, Eigen::RowMajor>> J(jacobian);
+  J.setZero();
+  J.block<2, 4>(0, 3) = 2.0 * okvis::kinematics::plus(quat).topLeftCorner<2, 4>();
+  J.col(6) = -1.0 * J.col(6);
+
+  return true;
+}
+
 // Computes the Jacobian from minimal space to naively overparameterised space as used by ceres.
-bool PoseLocalParameterization2d::liftJacobian(const double* x, double* jacobian) {
-  Eigen::Map<Eigen::Matrix<double, 2, 7, Eigen::RowMajor> > J_lift(jacobian);
+bool PoseManifold2d::liftJacobian(const double* x, double* jacobian) {
+  Eigen::Map<Eigen::Matrix<double, 2, 7, Eigen::RowMajor>> J_lift(jacobian);
   J_lift.setZero();
 
   const Eigen::Quaterniond q_inv(x[6], -x[3], -x[4], -x[5]);
@@ -398,9 +460,10 @@ bool PoseLocalParameterization2d::liftJacobian(const double* x, double* jacobian
 }
 
 // The jacobian of Plus(x, delta) w.r.t delta at delta = 0.
-bool PoseLocalParameterization2d::ComputeJacobian(const double* x, double* jacobian) const {
-  return plusJacobian(x, jacobian);
-}
+bool PoseManifold2d::PlusJacobian(const double* x, double* jacobian) const { return plusJacobian(x, jacobian); }
+
+// Compute the derivative of Minus(y, x) w.r.t y at y = x
+bool PoseManifold2d::MinusJacobian(const double* x, double* jacobian) const { return minusJacobian(x, jacobian); }
 
 }  // namespace ceres
 }  // namespace okvis
