@@ -610,6 +610,6 @@ BriefExtractor::BriefExtractor(const std::string& pattern_file) {
   m_brief.importPairs(x1, y1, x2, y2);
 }
 
-void Keyframe::setRelocalizationPCLCallback(const PointCloudCallback& pcl_callback) {
-  relocalization_pcl_callback_ = pcl_callback;
-}
+// void Keyframe::setRelocalizationPCLCallback(const PointCloudCallback& pcl_callback) {
+//   relocalization_pcl_callback_ = pcl_callback;
+// }
